**extracted worker and business from user**
Instead of trying to authenticate a Business now we authenticate Users in both cases, for workers and businesses.

Configuring two models for authentication would be overkill.

- [ ] Updated the create worker action to create both a worker and a user
- [ ] Update the create business action to create both a business and a user
- [ ] Update forms to bind accordingly.
- [ ] Removed the object binding, and replaced it with the objects values directly (added name and value attributes instead of th:field)
\
