--- src/main/java/com/codeup/sidework/daos/Businesses.java
@@ -1,8 +1,0 @@
-package com.codeup.sidework.daos;
-
-import com.codeup.sidework.models.Business;
-import org.springframework.data.repository.CrudRepository;
-
-public interface Businesses extends CrudRepository<Business, Long> {
-    Business findByUserName(String userName);
-}
