+++ src/main/java/com/codeup/sidework/daos/BusinessesRepository.java
@@ -4,5 +4,4 @@ import com.codeup.sidework.models.Business;
 import org.springframework.data.repository.CrudRepository;
 
 public interface BusinessesRepository extends CrudRepository<Business, Long> {
-    Business findByUserName(String userName);
 }
