+++ src/main/java/com/codeup/sidework/daos/WorkerRepository.java
@@ -1,0 +1,9 @@
+package com.codeup.sidework.daos;
+
+import com.codeup.sidework.models.Worker;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface WorkerRepository extends CrudRepository<Worker, Long> {
+}
