+++ src/main/resources/templates/users/register-mgmt.html
@@ -11,111 +11,55 @@
     <link rel="stylesheet" href="../../css/register-mgmt.css"/>
 </head>
 <body>
-<nav class="navbar navbar-default">
-    <div class="container-fluid">
-        <div class="navbar-header">
-            <a href="#" class="navbar-brand"><strong>sidework</strong></a>
-        </div>
-        <div>
-            <ul class="nav navbar-nav">
-                <li><a href="#">how it works</a></li>
-                <li><a href="#">español</a></li>
-                <li><a href="#">about us</a></li>
-            </ul>
-        </div>
-        <ul class="nav navbar-nav navbar-right">
-            <li><a href="#"><span class="glyphicon glyphicon-user" aria-hidden="true"></span> login</a></li>
-        </ul>
-    </div>
-</nav>
+<nav class="navbar navbar-default" th:replace="fragments/navbar :: mainNavbar"></nav>
 <br/>
 <div>
     <div class="container">
         <div class="pic"></div>
         <h1>register</h1>
-        <form th:action="@{/users/register-mgmt}" method="post" th:object="${business}">
+        <form th:action="@{/users/register-mgmt}" method="post">
             <div class="form-group">
                 <label for="username">username</label>
-                <input id="username" th:field="*{userName}" class="form-control" type="text"/>
+                <input id="username" name="username" th:value="${user.username}" class="form-control" type="text"/>
             </div>
             <div class="form-group">
                 <label for="email">email</label>
-                <input id="email" th:field="*{email}" class="form-control" type="email"/>
+                <input id="email" name="email" th:value="${user.email}" class="form-control" type="email"/>
             </div>
             <div class="form-group">
                 <label for="password">password</label>
-                <input id="password" th:field="*{password}" class="form-control" type="password"/>
+                <input id="password" name="password" class="form-control" type="password"/>
             </div>
             <div class="form-group">
                 <label for="confirm_password">confirm password</label>
-                <input id="confirm_password" th:field="*{password}" class="form-control" type="password"/>
-            </div>
-            <!--<form action="@{/registerbusiness}"></form>-->
-            <!--<div class="form-group">-->
-            <!--<label for="business_name">business name</label>-->
-            <!--<input id="business_name" th:field="*{username}" class="form-control" type="text"/>-->
-            <!--</div>-->
-            <!--<div class="alert alert-danger" role="alert">-->
-            <!--<strong>That username already exists</strong>-->
-            <!--</div>-->
-
-            <!--<div class="form-group">-->
-            <!--<label for="business_email">business email</label>-->
-            <!--<input id="business_email" th:field="*{email}" class="form-control" type="text"/>-->
-            <!--</div>-->
-            <!--<div class="alert alert-danger" role="alert">-->
-            <!--<strong>Not a valid email address</strong>-->
-            <!--</div>-->
-
-
-            <!--<div class="alert alert-danger" role="alert">-->
-            <!--<strong>Password doesn't match. Please try again.</strong>-->
-            <!--</div>-->
-
-
-            <div class="form-group">
-            <label for="business_phone_number">business phone number</label>
-            <input id="business_phone_number" th:field="*{businessPhone}" class="form-control" type="tel"/>
+                <input id="confirm_password" name="confirm_assword" class="form-control" type="password"/>
             </div>
 
             <div class="form-group">
-            <label for="business_website">business website</label>
-            <input id="business_website" th:field="*{website}" class="form-control" type="text"/>
+                <label for="business_phone_number">business name</label>
+                <input id="business_name" th:value="${business.businessName}" name="businessName" class="form-control" type="tel"/>
             </div>
 
             <div class="form-group">
-            <label for="business_address">address</label>
-            <input id="business_address" th:field="*{address}" class="form-control" type="textarea"/>
+                <label for="business_phone_number">business phone number</label>
+                <input id="business_phone_number" th:value="${business.businessPhone}" name="businessPhone" class="form-control" type="tel"/>
             </div>
 
             <div class="form-group">
-            <label for="info">info</label>
-            <input id="info" th:field="*{businessInfo}" class="form-control" type="textarea"/>
+                <label for="business_website">business website</label>
+                <input id="business_website" th:value="${business.website}" name="website" class="form-control" type="text"/>
             </div>
-            <!--<br/>-->
-            <!--<div>-->
-                <!--<h2>account manager</h2>-->
-                <!--<div class="panel-body">-->
-                    <!--<div class="panel panel-default">-->
-                        <!--&lt;!&ndash;<div class="panel-heading">currently employed?</div>&ndash;&gt;-->
-                        <!--<div class="form-group">-->
-                            <!--<label for="first_name">first name</label>-->
-                            <!--<input id="first_name" name="first_name" class="form-control" type="text"/>-->
-                            <!--<label for="last_name">last name</label>-->
-                            <!--<input id="last_name" name="last_name" class="form-control" type="text"/>-->
-                        <!--</div>-->
-                        <!--<div>-->
-                            <!--<label for="phone_number">phone number</label>-->
-                            <!--<input id="phone_number" name="phone_number" class="form-control" type="text"/>-->
-                        <!--</div>-->
-                        <!--<div>-->
-                            <!--<label for="position">position</label>-->
-                            <!--<input id="position" name="position" class="form-control" type="text"/>-->
-                        <!--</div>-->
-                    <!--</div>-->
-                <!--</div>-->
-            <!--</div>-->
-            <!--<br/>-->
+
+            <div class="form-group">
+                <label for="business_address">address</label>
+                <input id="business_address" th:value="${business.address}" name="address" class="form-control" type="textarea"/>
+            </div>
+
+            <div class="form-group">
+                <label for="info">info</label>
+                <input id="info" th:value="${business.businessInfo}" name="businessInfo" class="form-control" type="textarea"/>
+            </div>
+
             <div class="form-group">
                 <label><a href="#">please upload business owner verification letter</a></label>
             </div>
